+++ src/controller/controller-firebase.js
@@ -4,15 +4,6 @@ export const signIn = (email, password) => firebase.auth()
 export const logIn = (email, password) => firebase.auth()
   .signInWithEmailAndPassword(email, password);
 
-export const saveUsers = () => {
-  const user = firebase.auth().currentUser;
-  firebase.firestore().collection('users').doc(user.uid).set({
-    user: user.displayName,
-    avatar: user.photoURL,
-    uid: user.uid,
-    email: user.email,
-  });
-};
 
 export const googleLogin = () => {
   const provider = new firebase.auth.GoogleAuthProvider();
@@ -22,6 +13,10 @@ export const googleLogin = () => {
 
 firebase.initializeApp(firebaseConfig);
 let provider = new firebase.auth.GoogleAuthProvider(); 
+
+
+
+
